refactor(fuelConsumption): fix filename typo
Rename src/calcs/fuelConsumtion.ts -> fuelConsumption.ts (and its test
file). The optionKey 'economy<instance>' is unchanged, so persisted
configs continue to work without action.
