+++ src/calcs/fuelConsumption.ts
